+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/control-bar/control-bar.tsx
@@ -58,7 +58,7 @@ import { useWorkflowExecution } from '../../hooks/use-workflow-execution'
 import { DeploymentControls } from './components/deployment-controls/deployment-controls'
 import { ExportControls } from './components/export-controls/export-controls'
 import { HistoryDropdownItem } from './components/history-dropdown-item/history-dropdown-item'
-import { ImportControls } from './components/import-controls/import-controls'
+
 import { MarketplaceModal } from './components/marketplace-modal/marketplace-modal'
 import { NotificationDropdownItem } from './components/notification-dropdown-item/notification-dropdown-item'
 import { UserAvatarStack } from './components/user-avatar-stack/user-avatar-stack'
@@ -1289,7 +1289,7 @@ export function ControlBar({ hasValidationErrors = false }: ControlBarProps) {
         {renderDuplicateButton()}
         {renderAutoLayoutButton()}
         {renderDebugModeToggle()}
-        <ImportControls disabled={!userPermissions.canEdit} />
+
         <ExportControls disabled={!userPermissions.canRead} />
         {/* <WorkflowTextEditorModal disabled={!userPermissions.canEdit} /> */}
         {/* {renderPublishButton()} */}
