+++ apps/sim/app/workspace/[workspaceId]/w/components/sidebar/components/create-menu/create-menu.tsx
@@ -1,16 +1,19 @@
 'use client'
 
-import { useState } from 'react'
+import { useState, useRef } from 'react'
 import { logger } from '@sentry/nextjs'
-import { File, Folder, Plus } from 'lucide-react'
+import { ChevronRight, File, Folder, Plus, Upload } from 'lucide-react'
 import { useParams } from 'next/navigation'
 import { Button } from '@/components/ui/button'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
+import { Separator } from '@/components/ui/separator'
 import { cn } from '@/lib/utils'
 import { useFolderStore } from '@/stores/folders/store'
+import { useUserPermissionsContext } from '@/app/workspace/[workspaceId]/w/components/providers/workspace-permissions-provider'
+import { ImportControls, ImportControlsRef } from './import-controls'
 
 interface CreateMenuProps {
   onCreateWorkflow: (folderId?: string) => void
@@ -27,10 +30,15 @@ export function CreateMenu({
   const [folderName, setFolderName] = useState('')
   const [isCreating, setIsCreating] = useState(false)
   const [isHoverOpen, setIsHoverOpen] = useState(false)
+  const [isImportSubmenuOpen, setIsImportSubmenuOpen] = useState(false)
 
   const params = useParams()
   const workspaceId = params.workspaceId as string
   const { createFolder } = useFolderStore()
+  const userPermissions = useUserPermissionsContext()
+
+  // Ref for the file input that will be used by ImportControls
+  const importControlsRef = useRef<ImportControlsRef>(null)
 
   const handleCreateWorkflow = () => {
     setIsHoverOpen(false)
@@ -40,6 +48,13 @@ export function CreateMenu({
   const handleCreateFolder = () => {
     setIsHoverOpen(false)
     setShowFolderDialog(true)
+  }
+
+  const handleUploadYaml = () => {
+    setIsHoverOpen(false)
+    setIsImportSubmenuOpen(false)
+    // Trigger the file upload from ImportControls component
+    importControlsRef.current?.triggerFileUpload()
   }
 
   const handleFolderSubmit = async (e: React.FormEvent) => {
@@ -99,7 +114,7 @@ export function CreateMenu({
             'data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2',
             'z-50 animate-in overflow-hidden rounded-md border bg-popover p-1 text-popover-foreground shadow-md',
             'data-[state=closed]:animate-out',
-            'w-40'
+            'w-48'
           )}
           onMouseEnter={() => setIsHoverOpen(true)}
           onMouseLeave={() => setIsHoverOpen(false)}
@@ -119,6 +134,7 @@ export function CreateMenu({
             <File className='h-4 w-4' />
             {isCreatingWorkflow ? 'Creating...' : 'New Workflow'}
           </button>
+          
           <button
             className='flex w-full cursor-default select-none items-center gap-2 rounded-sm px-2 py-1.5 text-sm outline-none transition-colors hover:bg-accent hover:text-accent-foreground'
             onClick={handleCreateFolder}
@@ -126,8 +142,62 @@ export function CreateMenu({
             <Folder className='h-4 w-4' />
             New Folder
           </button>
+
+          {userPermissions.canEdit && (
+            <>
+              <Separator className='my-1' />
+              
+              <Popover open={isImportSubmenuOpen} onOpenChange={setIsImportSubmenuOpen}>
+                <PopoverTrigger asChild>
+                  <button
+                    className='flex w-full cursor-default select-none items-center justify-between rounded-sm px-2 py-1.5 text-sm outline-none transition-colors hover:bg-accent hover:text-accent-foreground'
+                    onMouseEnter={() => setIsImportSubmenuOpen(true)}
+                  >
+                    <div className='flex items-center gap-2'>
+                      <Upload className='h-4 w-4' />
+                      <span>Import Workflow</span>
+                    </div>
+                    <ChevronRight className='h-3 w-3' />
+                  </button>
+                </PopoverTrigger>
+                <PopoverContent
+                  side='right'
+                  align='start'
+                  sideOffset={4}
+                  className='w-48 p-1'
+                  onMouseEnter={() => setIsImportSubmenuOpen(true)}
+                  onMouseLeave={() => setIsImportSubmenuOpen(false)}
+                  onOpenAutoFocus={(e) => e.preventDefault()}
+                  onCloseAutoFocus={(e) => e.preventDefault()}
+                >
+                  <button
+                    className='flex w-full cursor-default select-none items-center gap-2 rounded-sm px-2 py-1.5 text-sm outline-none transition-colors hover:bg-accent hover:text-accent-foreground'
+                    onClick={handleUploadYaml}
+                  >
+                    <Upload className='h-4 w-4' />
+                    <div className='flex flex-col items-start'>
+                      <span>YAML</span>
+                      <span className='text-muted-foreground text-xs'>
+                        .yaml or .yml
+                      </span>
+                    </div>
+                  </button>
+                </PopoverContent>
+              </Popover>
+            </>
+          )}
         </PopoverContent>
       </Popover>
+
+      {/* Import Controls Component - handles all import functionality */}
+      <ImportControls 
+        ref={importControlsRef}
+        disabled={!userPermissions.canEdit}
+        onClose={() => {
+          setIsHoverOpen(false)
+          setIsImportSubmenuOpen(false)
+        }}
+      />
 
       {/* Folder creation dialog */}
       <Dialog open={showFolderDialog} onOpenChange={setShowFolderDialog}>
